+++ client/app/store.js
@@ -2,7 +2,7 @@ import { configureStore } from "@reduxjs/toolkit";
 import logger from "redux-logger";
 import authReducer from "../features/auth/authSlice";
 import productsReducer from "../features/products/productsSlice";
-import singleProductReducer from "../features/SingleProducts/singleProductSlice";
+import singleProductReducer from "../features/singleProduct/singleProductSlice";
 
 const store = configureStore({
   reducer: {
