+++ client/features/singleProduct/Info.js
