+++ client/features/singleProduct/MainImage.js
